notifier/requirements.txt
python-telegram-bot>=13.0
pika>=1.2.0
